requirements.txt
scipy==1.4.1
numpy==1.18.1
pyjnius==1.2.1
sacred==0.8.0
pandas==1.0.3
matplotlib==3.1.2
func_timeout==4.3.5
